+++ Goulib/__init__.py
@@ -9,4 +9,4 @@ for various projects
 Project location : https://github.com/goulu/goulib
 """
 
-__version__ = '1.6.0'
+__version__ = '1.6.1'