+++ CHANGES.rst
@@ -1,3 +1,8 @@
+v1.6.1, 2013-12-11
+* :meth:`table.Table.html` now returns a string (no more dependent on :mod:`markup`)
+* :class:`table.Table` can be init'ed from an HTML file using `lxml <http://pypi.python.org/pypi/lxml/>`
+* added support for nose coverage tests
+  
 v1.6.0, 2013-11-25
 * added :mod:`geom`
 * :mod:`dxf2img` improved and added factory method to generate geom objects from dxf